clean up CubeGenerator class
